[36] move constant and add export
